Putting the if prefix commands into their own files
